# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "iGovTT AI Tools",
        "version": "1.0.0",
        "description": "Tools used by tawk.to AI Assist to retrieve iGovTT services and create appointment links."
    },
    "servers": [
        {
            "url": "https://tawkto-eapp.onrender.com"
        }
    ],
    "paths": {
        "/api/catalog/services": {
            "get": {
                "operationId": "listServices",
                "summary": "List iGovTT appointment services",
                "description": "Always use this tool when a visitor asks to list services, asks what appointment services are available, asks what services can be booked, or asks for available iGovTT services. Do not answer from general knowledge. Only answer using the services returned by this API.",
                "responses": {
                    "200": {
                        "description": "List of available appointment services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "question": {
                                            "type": "string"
                                        },
                                        "services": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalog/appointment-deeplink": {
            "post": {
                "operationId": "createAppointmentDeepLink",
                "summary": "Create an appointment deeplink",
                "description": "Use this tool immediately when the visitor provides both serviceId and branchId and wants an appointment booking link. Do not ask for branch name if branchId is provided. After the tool returns, reply with the appointmentUrl.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "serviceId",
                                    "branchId"
                                ],
                                "properties": {
                                    "serviceId": {
                                        "type": "string",
                                        "description": "The numeric service ID supplied by the visitor. Use the exact value given by the visitor."
                                    },
                                    "branchId": {
                                        "type": "string",
                                        "description": "The numeric branch ID supplied by the visitor. Use the exact value given by the visitor."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment link created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "appointmentUrl": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalog/branches": {
            "post": {
                "operationId": "listBranches",
                "summary": "List branches for a selected appointment service",
                "description": "Use this tool after the visitor selects an appointment service. Pass the serviceId returned by listServices. Show branch names and addresses to the visitor, but do not expose branch IDs. Retain the selected branch ID for later appointment booking.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "serviceId"
                                ],
                                "properties": {
                                    "serviceId": {
                                        "type": "string",
                                        "description": "The service ID returned by the listServices tool for the service selected by the visitor."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branches available for the selected service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "question": {
                                            "type": "string"
                                        },
                                        "serviceId": {
                                            "type": "string"
                                        },
                                        "branches": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalog/available-dates": {
            "post": {
                "operationId": "listAvailableDates",
                "summary": "List available appointment dates",
                "description": "Use this tool after the visitor selects a service and branch. Pass the selected serviceId and branchId. Show the date labels to the visitor, but retain the date value in YYYY-MM-DD format for the next tool call.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "serviceId",
                                    "branchId"
                                ],
                                "properties": {
                                    "serviceId": {
                                        "type": "string",
                                        "description": "The service ID returned by listServices."
                                    },
                                    "branchId": {
                                        "type": "string",
                                        "description": "The branch ID returned by listBranches."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Available dates for the selected service and branch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "question": {
                                            "type": "string"
                                        },
                                        "serviceId": {
                                            "type": "string"
                                        },
                                        "branchId": {
                                            "type": "string"
                                        },
                                        "dates": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}